Restructure tests to allow for more complex testing, and vscode to ignore file
