+++ .gitignore
@@ -20,3 +20,4 @@ test/screenshots/
 .local
 .yarnrc
 .pki
+.vscode
